Use a client package to disambiguate
* Import paths remain the same and are aliased in the main __init__.py
* This moves to use a nested client package to disambiguate the purpose
and allow for future flexibility to implement an oauth2 server.
* Fixes an error when callback handlers are not provided.
